requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   google-genai
    #   httpx
boto3==1.42.10
    # via -r requirements.in
botocore==1.42.11
    # via
    #   boto3
    #   s3transfer
cachetools==6.2.4
    # via google-auth
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
duckdb==1.4.3
    # via -r requirements.in
google-auth==2.45.0
    # via google-genai
google-genai==1.33.0
    # via -r requirements.in
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via google-genai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.5
    # via google-genai
pydantic-core==2.41.5
    # via pydantic
python-dateutil==2.9.0.post0
    # via botocore
requests==2.32.4
    # via
    #   -r requirements.in
    #   google-genai
rsa==4.9.1
    # via google-auth
s3transfer==0.16.0
    # via boto3
six==1.17.0
    # via python-dateutil
tenacity==9.1.2
    # via google-genai
typing-extensions==4.15.0
    # via
    #   google-genai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.2
    # via
    #   botocore
    #   requests
websockets==15.0.1
    # via google-genai
